Remove dead and move around some code
