+++ soroban-env-host/src/host/error.rs
@@ -1,15 +1,11 @@
 use crate::{
     budget::AsBudget,
     events::Events,
-    xdr::{self, ScError},
-    EnvBase, Error, Host,
+    xdr::{self, Hash, ScAddress, ScError, ScErrorCode, ScErrorType},
+    ConversionError, EnvBase, Error, Host, TryFromVal, U32Val, Val,
 };
 use backtrace::{Backtrace, BacktraceFrame};
 use core::fmt::Debug;
-use soroban_env_common::{
-    xdr::{ScErrorCode, ScErrorType},
-    ConversionError, TryFromVal, U32Val, Val,
-};
 use std::{
     cell::{Ref, RefCell, RefMut},
     ops::DerefMut,
@@ -325,6 +321,77 @@ impl Host {
                 self.error(e.into(), "", &[])
             }
         })
+    }
+
+    pub(crate) fn decorate_contract_data_storage_error(
+        &self,
+        err: HostError,
+        key: Val,
+    ) -> HostError {
+        if !err.error.is_type(ScErrorType::Storage) {
+            return err;
+        }
+        if err.error.is_code(ScErrorCode::ExceededLimit) {
+            return self.err(
+                ScErrorType::Storage,
+                ScErrorCode::ExceededLimit,
+                "trying to access contract storage key outside of the correct footprint",
+                &[key],
+            );
+        }
+        if err.error.is_code(ScErrorCode::MissingValue) {
+            return self.err(
+                ScErrorType::Storage,
+                ScErrorCode::MissingValue,
+                "trying to get non-existing value for contract storage key",
+                &[key],
+            );
+        }
+        err
+    }
+
+    pub(crate) fn decorate_contract_instance_storage_error(
+        &self,
+        err: HostError,
+        contract_id: &Hash,
+    ) -> HostError {
+        if err.error.is_type(ScErrorType::Storage) && err.error.is_code(ScErrorCode::ExceededLimit)
+        {
+            return self.err(
+                ScErrorType::Storage,
+                ScErrorCode::ExceededLimit,
+                "trying to access contract instance key outside of the correct footprint",
+                // No need for metered clone here as we are on the unrecoverable
+                // error path.
+                &[self
+                    .add_host_object(ScAddress::Contract(contract_id.clone()))
+                    .map(|a| a.into())
+                    .unwrap_or(Val::VOID.into())],
+            );
+        }
+        err
+    }
+
+    pub(crate) fn decorate_contract_code_storage_error(
+        &self,
+        err: HostError,
+        wasm_hash: &Hash,
+    ) -> HostError {
+        if err.error.is_type(ScErrorType::Storage) && err.error.is_code(ScErrorCode::ExceededLimit)
+        {
+            return self.err(
+                ScErrorType::Storage,
+                ScErrorCode::ExceededLimit,
+                "trying to access contract code key outside of the correct footprint",
+                // No need for metered clone here as we are on the unrecoverable
+                // error path.
+                &[self
+                    .add_host_object(self.scbytes_from_hash(wasm_hash).unwrap_or_default())
+                    .map(|a| a.into())
+                    .unwrap_or(Val::VOID.into())],
+            );
+        }
+        err
     }
 }
